+++ scripts/devtools/tpl-manager.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-import argparse, os, sys, re, shutil
+import argparse, os, sys, re, shutil, glob, time
 
 sys.path.append(os.path.join(os.path.dirname(__file__), ".."))
 from spheralutils import sexe
@@ -46,6 +46,8 @@ class SpheralTPL:
                            choices=list(dev_specs.keys()))
         parser.add_argument("--show-specs", action="store_true",
                             help="Show the specs for the current environment and stop.")
+        parser.add_argument("--show-info", action="store_true",
+                            help="Show the Spheral Spack info, including dependencies and variants.")
         parser.add_argument("--add-spec", action="store_true",
                             help="Set this flag to add the --spec to the environment.")
         parser.add_argument("--spack-dir", type=str,
@@ -109,41 +111,56 @@ class SpheralTPL:
         spack_dir = os.path.join(tpl_root, "spack")
         if (not self.args.skip_init):
             if (not os.path.exists(spack_dir)):
-                sexe(f"git init {spack_dir}")
-                sexe(f"git -C {spack_dir} remote add origin {self.args.spack_url}")
+                sexe(f"git -C {tpl_root} clone --depth=2 {self.args.spack_url}")
+            # Check commit hash of Spack repo
+            cur_hash = sexe(f"git -C {spack_dir} rev-parse HEAD", ret_output=True, echo=False).strip()
+            if (cur_hash != spack_commit):
                 sexe(f"git -C {spack_dir} fetch --depth=2 origin {spack_commit}")
                 sexe(f"git -C {spack_dir} checkout FETCH_HEAD")
-            else:
-                # Check commit hash of Spack repo
-                cur_hash = sexe(f"git -C {spack_dir} rev-parse HEAD", ret_output=True, echo=False).strip()
-                if (cur_hash != spack_commit):
-                    sexe(f"git -C {spack_dir} fetch --depth=2 origin {spack_commit}")
-                    sexe(f"git -C {spack_dir} checkout FETCH_HEAD")
+            uber_env_trash = os.path.join(spack_dir, "etc/spack/defaults/upstreams.yaml")
+            if (self.args.clean and os.path.exists(uber_env_trash)):
+                sexe(f"git -C {spack_dir} clean -df")
         self.add_spack_paths(spack_dir)
-        if (self.args.clean):
-            sexe("rm -rf ~/.spack")
-            sexe(f"git -C {spack_dir} clean -df")
 
-    def find_spack_package(self, package_names):
+    def check_lock_file(self):
+        "Check if any files in scripts/spack are newer than the spack.lock file"
+        spack_lock = os.path.join(self.env_dir, "spack.lock")
+        if (not os.path.exists(spack_lock)):
+            return False
+        lock_time = os.path.getmtime(spack_lock)
+        for package, path in package_dirs.items():
+            internal_spack_dir = os.path.join(os.path.abspath(get_config_dir(path)), "**")
+            sp_files = glob.glob(internal_spack_dir, recursive=True)
+            ctimes = [os.path.getmtime(ff) for ff in sp_files\
+                      if os.path.isfile(ff) and "__pycache__" not in ff]
+            if (any(x > lock_time for x in ctimes)):
+                return True
+        return False
+
+    def find_spack_package(self, package_name, req=True):
         """
         Find if the package name/s is already in the environment.
         If any one of them are found, return the name of that package.
         If not, run spack external find and raise an error if none are found.
         """
         cur_packages = spack.config.get("packages")
-        if (type(package_names) is not list):
-            package_names = [package_names]
-        for i in package_names:
-            if (i in cur_packages):
-                return i
+        if (package_name in cur_packages):
+            return True
         ext_cmd = SpackCommand("external")
-        find_out = ext_cmd("find", "--not-buildable", *package_names)
-        for i in package_names:
-            if (i in find_out):
-                return i
-        raise Exception(f"System install of {package_names} not found. "+\
-                        "If software is installed, add location to $PATH "+\
-                        "environment variable. Otherwise, install package.")
+        if (req):
+            find_out = ext_cmd("find", "--not-buildable", package_name)
+        else:
+            find_out = ext_cmd("find", package_name)
+        if (package_name in find_out):
+            return True
+        return None
+
+    def find_all_spack_packages(self, packages, req=True):
+        for i in packages:
+            if (not self.find_spack_package(i, req) and req):
+                raise Exception(f"System install of {i} not found. "+\
+                                "If software is installed, add location to $PATH "+\
+                                "environment variable. Otherwise, install package.")
 
     def modify_env_file(self, env_file, mod_func):
         "Modify the spack.yaml file"
@@ -158,18 +175,22 @@ class SpheralTPL:
         with open(env_file, 'w') as ff:
             spack_yaml.dump(loader, ff)
 
-    def custom_spack_env(self, env_dir, env_name):
+    def custom_spack_env(self, env_name):
         "Use/create a custom Spack environment"
         from spack import environment
         if (not self.args.spec):
-            raise Exception("Must supply a --spec for a custom environment")
-
-        if (not os.path.exists(os.path.join(env_dir, "spack.yaml"))):
+            raise Exception("Must supply a --spec for a custom environment (IE --spec spheral+mpi%gcc)")
+        env_file = os.path.join(self.env_dir, "spack.yaml")
+        if (not os.path.exists(env_file)):
             # Create a new environment
             env_cmd = SpackCommand("env")
-            env_cmd("create", "--without-view", "-d", env_dir)
-
-        self.spack_env = environment.Environment(env_dir)
+            env_cmd("create", "--without-view", "-d", self.env_dir)
+            def set_concretize(loader):
+                new_dict = {"concretizer": {"unify": False}}
+                loader["spack"].update(new_dict)
+                return loader
+            self.modify_env_file(env_file, set_concretize)
+        self.spack_env = environment.Environment(self.env_dir)
         environment.activate(self.spack_env)
         # Get all the Spack commands
         repo_cmd = SpackCommand("repo")
@@ -185,30 +206,51 @@ class SpheralTPL:
                 repo_cmd("add", f"{repo_path}") # spack repo add <repo_path>
             dev_path = os.path.abspath(path)
             dev_cmd("-p", dev_path, f"{package}@=develop") # spack develop <package>@=develop
-
         comp_cmd("find") # spack compiler find
         ext_cmd("find") # spack external find
+        provider_dict = {}
 
         # req_packages are packages we refuse to let spack build
-        # If they aren't found, an error will be thrown telling the user
-        # to install the package or ensure an install is in $PATH
-        req_packages = ["python", "perl"]
-        for i in req_packages:
-            self.find_spack_package(i)
+        # If they aren't found on the system, an error will be thrown
+        # telling the user to install the package or ensure an install is in $PATH
+        self.find_all_spack_packages(["python", "perl"])
         # Look for MPI installs
         if (not spack.spec.Spec(self.args.spec).satisfies("~mpi")):
-            mpi_packages = ["mpich", "openmpi"]
-            mpi_pack = self.find_spack_package(mpi_packages)
-            if (f"^{mpi_pack}" not in self.args.spec):
-                self.args.spec += f"^{mpi_pack}"
+            found_mpi = False
+            for i in ["openmpi", "mpich"]:
+                if (self.find_spack_package(i)):
+                    found_mpi = True
+                    provider_dict.update({"mpi": [i]})
+                    break
+            if (not found_mpi):
+                raise Exception(f"System MPI install not found. "+\
+                                "If software is installed, add location to $PATH "+\
+                                "environment variable. Otherwise, install package.")
+        # opt_packages are packages we would like to find on the system
+        # but are ok with Spack installing if they are not found
+        opt_packages = ["hdf5", "ncurses"]
+        for i in opt_packages:
+            self.find_spack_package(i, req=False)
+        # Hard-coding providers for these packages vastly improves TPL system
+        provider_dict.update({"zlib-api": ["zlib"],
+                              "blas": ["openblas"],
+                              "lapack": ["openblas"]})
         # Always add the spec for a custom environment
+        if (provider_dict):
+            self.config_env_providers(provider_dict)
         self.args.add_spec = True
 
-    def remove_upstream(self, env_dir):
+    def config_env_providers(self, config_dict):
+        env_file = os.path.join(self.env_dir, "spack.yaml")
+        def set_providers(loader):
+            new_dict = {"all": {"providers": config_dict}}
+            loader["spack"]["packages"].update(new_dict)
+            return loader
+        self.modify_env_file(env_file, set_providers)
+
+    def remove_upstream(self):
         "Modify the spack.yaml to remove the upstream"
-        # TODO: Currently, Spack has no other way to
-        # to remove an include: line from an environment
-        # than to directly change the spack.yaml file
+        # TODO: Ideally use spack config command to achieve this
         # Remove upstream.yaml or upstream entry
         def do_remove(loader):
             if ("upstreams" in loader["spack"]):
@@ -221,8 +263,8 @@ class SpheralTPL:
 
         # Copy spack.yaml to origspack.yaml and overwrite spack.yaml
         # with upstreams removed
-        env_file = os.path.join(env_dir, "spack.yaml")
-        shutil.copyfile(env_file, os.path.join(env_dir, "origspack.yaml"))
+        env_file = os.path.join(self.env_dir, "spack.yaml")
+        shutil.copyfile(env_file, os.path.join(self.env_dir, "origspack.yaml"))
         self.modify_env_file(env_file, do_remove)
 
     def activate_spack_env(self):
@@ -234,19 +276,19 @@ class SpheralTPL:
             default_env = "dev_pkg"
         if default_env and os.path.exists(os.path.join(config_env_dir, default_env)):
             # For LC systems
-            env_dir = os.path.join(config_env_dir, default_env)
-            print(f"Activating Spack environment in {env_dir}")
+            self.env_dir = os.path.join(config_env_dir, default_env)
+            print(f"Activating Spack environment in {self.env_dir}")
             if self.args.no_upstream:
-                self.remove_upstream(env_dir)
+                self.remove_upstream()
             from spack import environment
-            self.spack_env = environment.Environment(env_dir)
+            self.spack_env = environment.Environment(self.env_dir)
             environment.activate(self.spack_env)
         else:
             # Otherwise, check if environment has been created
             arch_cmd = SpackCommand("arch")
             env_name = arch_cmd().strip()
-            env_dir = os.path.join(config_env_dir, env_name)
-            self.custom_spack_env(env_dir, env_name)
+            self.env_dir = os.path.join(config_env_dir, env_name)
+            self.custom_spack_env(env_name)
 
     def concretize_spec(self, check_spec):
         "Concretize the spec"
@@ -255,9 +297,14 @@ class SpheralTPL:
             if (self.args.add_spec):
                 add_cmd = SpackCommand("add")
                 add_cmd(self.args.spec)
-        print("Concretizing environment")
         conc_cmd = SpackCommand("concretize")
-        conc_cmd("-U", "-f")
+        conc_args = ["-U"]
+        if (self.args.clean):
+            conc_args.append("-f")
+            print("Cleaning and concretizing environment")
+        else:
+            print("Concretizing environment")
+        conc_cmd(*conc_args)
         if (check_spec):
             matches = self.spack_env.matching_spec(self.spack_spec)
             if (not matches):
@@ -296,21 +343,23 @@ class SpheralTPL:
             spec_cmd = SpackCommand("spec")
             print(f"Running spack spec -IL {spec}")
             spec_cmd("-IL", spec)
-        install_args = ["-u", "initconfig"]
+        install_args = ["-u", "initconfig", "--fail-fast"]
         if (self.args.dev_pkg):
             # Spec is provided so assumes we are building from a buildcache
             install_args.extend(["--use-buildcache", "package:never,dependencies:only", "--no-check-signature"])
         self.do_install(install_args, spec)
-        print(f"Created {host_config_file}")
-
         if (self.args.ci_run):
             shutil.copyfile(host_config_file, "gitlab.cmake")
-
+            host_config_file = "gitlab.cmake"
         if (mod_host_config):
             # Apply --id and bring back original host config file
             new_name = host_config_file.replace(".cmake", f"{self.args.id}.cmake")
             os.rename(host_config_file, new_name)
             os.rename("orig"+host_config_file, host_config_file)
+            host_config_file = new_name
+        print(f"Created {host_config_file}. To configure, run:\n")
+        scr_path = os.path.relpath(package_dirs['spheral'], base_dir)
+        print(f"{scr_path}/scripts/devtools/host-config-build.py --host-config {host_config_file}\n")
 
     def __init__(self):
         self.parse_args()
@@ -322,6 +371,13 @@ class SpheralTPL:
             find_cmd = SpackCommand("find")
             find_cmd("-r")
             sys.exit(0)
+        if (self.args.show_info):
+            info_cmd = SpackCommand("info")
+            info_cmd("spheral")
+            sys.exit(0)
+        # Check if any files in scripts/spack are newer than the spack.lock
+        if (not self.args.clean):
+            self.args.clean = self.check_lock_file()
         if (self.args.spec):
             # If --spec is given, install TPLs and create host config file
             self.concretize_spec(check_spec=True)
@@ -330,7 +386,7 @@ class SpheralTPL:
             # Concretize the current environment
             self.concretize_spec(check_spec=False)
             # No spec is given, install TPLs for all env specs
-            install_args = ["--only", "dependencies"]
+            install_args = ["-u", "initconfig", "--fail-fast"]
             self.do_install(install_args, package_name)
 
         # Undo any file changes we made to spack.yaml
@@ -338,6 +394,12 @@ class SpheralTPL:
         if (self.args.no_upstream and os.path.exists(orig_file)):
             # Revert env file if it was modified
             os.rename(orig_file, os.path.join(self.spack_env.path, "spack.yaml"))
+        # Remove symbolic directory created by Spack
+        print("Removing Spack symbolic build directories")
+        build_dirs = glob.glob("build-*")
+        for i in build_dirs:
+            if (os.path.islink(i)):
+                os.unlink(i)
 
 if __name__=="__main__":
     spheral_tpl = SpheralTPL()