+++ scripts/spack/packages/spheral/package.py
@@ -73,20 +73,21 @@ class Spheral(CachedCMakePackage, CudaPackage, ROCmPackage):
     with when('~rocm') or when('~cuda'):
         depends_on('axom +shared', type='build')
 
-    with when("+caliper"):
-        depends_on("caliper@2.11 ~shared +adiak +gotcha ~libdw ~papi ~libunwind +pic", type="build")
-        depends_on("caliper+mpi", type="build", when="+mpi")
-        depends_on("caliper~mpi", type="build", when="~mpi")
+    with when('+caliper'):
+        depends_on('caliper@2.11 ~shared +adiak +gotcha ~libdw ~papi ~libunwind cppflags="-fPIC"', type='build')
+        depends_on('caliper+mpi', type='build', when='+mpi')
+        depends_on('caliper~mpi', type='build', when='~mpi')
 
-    depends_on("raja@2024.02.0", type="build")
+    depends_on('raja@2024.02.0', type='build')
 
     depends_on('opensubdiv@3.4.3+pic', type='build', when="+opensubdiv")
 
-    depends_on('polytope +python', type='build')
+    depends_on('polytope +python', type='build', when="+python")
+    depends_on('polytope ~python', type='build', when="~python")
 
-    depends_on('sundials@7.0.0 ~shared cxxstd=17', type='build', when='+sundials')
+    depends_on('sundials@7.0.0 ~shared cxxstd=17 cppflags="-fPIC"', type='build', when='+sundials')
 
-    depends_on("leos@8.4.2", type="build", when="+leos")
+    depends_on('leos@8.4.2', type='build', when='+leos')
 
     # Forward MPI Variants
     mpi_tpl_list = ["hdf5", "conduit", "axom", "adiak~shared"]
@@ -106,6 +107,7 @@ class Spheral(CachedCMakePackage, CudaPackage, ROCmPackage):
     # Conflicts
     # -------------------------------------------------------------------------
     conflicts("+cuda", when="+rocm")
+    conflicts("%pgi")
 
     def _get_sys_type(self, spec):
         sys_type = spec.architecture
@@ -205,7 +207,6 @@ class Spheral(CachedCMakePackage, CudaPackage, ROCmPackage):
         # TPL locations
         if (spec.satisfies("+caliper")):
             entries.append(cmake_cache_path('caliper_DIR', spec['caliper'].prefix))
-            #entries.append(cmake_cache_option('ENABLE_TIMER', True))
 
         entries.append(cmake_cache_path('adiak_DIR', spec['adiak'].prefix))
 
@@ -232,10 +233,10 @@ class Spheral(CachedCMakePackage, CudaPackage, ROCmPackage):
 
         if (spec.satisfies("+opensubdiv")):
             entries.append(cmake_cache_path('opensubdiv_DIR', spec['opensubdiv'].prefix))
-            entries.append(cmake_cache_path('ENABLE_OPENSUBDIV', True))
+            entries.append(cmake_cache_option('ENABLE_OPENSUBDIV', True))
 
         if (spec.satisfies("~network")):
-            entries.append(cmake_cache_path('SPHERAL_NETWORK_CONNECTED', False))
+            entries.append(cmake_cache_option('SPHERAL_NETWORK_CONNECTED', False))
 
         entries.append(cmake_cache_path('polytope_DIR', spec['polytope'].prefix))